refactoring: moved existing test cases for methods of classes into its own file
